move contents of overlay/tests.rs to overlay.rs
